projeto/back/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.3
scipy==1.15.2
